chore: remove logEntry object from errors
also had to move stuff around to get rid of circular dependencies
